Update san_francisco_housing.ipynb
Answered the Questions: Did any year experience a drop in the average sale price per square foot compared to the previous year? Answer: The only slight drop in price was between 2010 to 2011 in the average sale price per square foot compared to the previous year. The majority has been a gradual increase.

If so, did the gross rent increase or decrease during that year?
The gross rent increased every year between 2010 and 2016. In the year that sale price per sqft. decreased, the gross rent stabalized slightly. But we do not see this same trend from 2014 to 2015 when the gross rent stabalized again. More research would need to be done to explain this.