Update san_francisco_housing.ipynb
Answered the Questions: How does the trend in rental income growth compare to the trend in sales prices? Does this same trend hold true for all the neighborhoods across San Francisco?

Answer: As rental income growth increases, so does the trend in sales price. This not true for all neighborhoods from what I can see. Some neighborhoods have a variety of rents and sales prices.

What insights can you share with your company about the potential one-click, buy-and-rent strategy that they're pursuing? Do neighborhoods exist that you would suggest for investment, and why?

Answer: Yes, there are some neighborhoods that have a low sale price and a high gross rent that would make renting it out proffitable. A good example of this is Silver Terrace, where the sale price per sq.ft is 170.29 and the gross rent is 3528.